+++ src/main/java/com/jcraft/jsch/ChannelSftp.java
@@ -604,8 +604,10 @@ public class ChannelSftp extends ChannelSession {
                 int _ackid = ackid[0];
                 if (startid > _ackid || _ackid > seq - 1) {
                   if (_ackid == seq) {
-                    System.err.println(
-                        "ack error: startid=" + startid + " seq=" + seq + " _ackid=" + _ackid);
+                    if (getSession().getLogger().isEnabled(Logger.ERROR)) {
+                      getSession().getLogger().log(Logger.ERROR,
+                          "ack error: startid=" + startid + " seq=" + seq + " _ackid=" + _ackid);
+                    }
                   } else {
                     throw new SftpException(SSH_FX_FAILURE,
                         "ack error: startid=" + startid + " seq=" + seq + " _ackid=" + _ackid);