+++ src/main/java/com/jcraft/jsch/jce/SHA256.java
@@ -40,11 +40,7 @@ public class SHA256 implements HASH {
 
   @Override
   public void init() throws Exception {
-    try {
-      md = MessageDigest.getInstance("SHA-256");
-    } catch (Exception e) {
-      System.err.println(e);
-    }
+    md = MessageDigest.getInstance("SHA-256");
   }
 
   @Override